# lib/db/src/schema/botConfig.ts
import {
  pgTable,
  text,
  serial,
  boolean,
  integer,
  numeric,
  timestamp,
} from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const botConfigTable = pgTable("bot_config", {
  id: serial("id").primaryKey(),
  isRunning: boolean("is_running").notNull().default(false),
  checkIntervalSeconds: integer("check_interval_seconds").notNull().default(30),
  maxConcurrentBets: integer("max_concurrent_bets").notNull().default(5),
  dailyLossLimit: numeric("daily_loss_limit", { precision: 10, scale: 2 }).notNull().default("100.00"),
  dailyProfitTarget: numeric("daily_profit_target", { precision: 10, scale: 2 }),
  enabledEventTypes: text("enabled_event_types").notNull().default("1,2,4"),
  paperTradingMode: boolean("paper_trading_mode").notNull().default(true),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertBotConfigSchema = createInsertSchema(botConfigTable).omit({
  id: true,
  updatedAt: true,
});
export type InsertBotConfig = z.infer<typeof insertBotConfigSchema>;
export type BotConfig = typeof botConfigTable.$inferSelect;